backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
numpy==1.26.4
scipy==1.12.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
python-multipart==0.0.9

